// src/app/modules/accounting/models/verify-loan-application.model.ts
export class SelectedLoanGetModel{
    AccFlag: number;
    AccTypeDesc: string;
    AtyClass: number;
    InstallmentAmount: number;
    LastInstallmentAmount: number;
    NoOfInstallment: number;
    PreviousPerformence: string;
    SacnAmount: number;
    StepOfLoan: number;
    TotalDeposit: number;
}